Add MVVM
