--- app/src/main/java/com/ds_create/bulletinboard/data/Ad.kt
@@ -1,15 +1,0 @@
-package com.ds_create.bulletinboard.data
-
-data class Ad(
-    val country: String? = null,
-    val city: String? = null,
-    val tel: String? = null,
-    val index: String? = null,
-    val withSend: String? = null,
-    val category: String? = null,
-    val title: String? = null,
-    val price: String? = null,
-    val description: String? = null,
-    val key: String? = null,
-    val uid: String? = null
-)
