--- app/src/main/java/com/ds_create/bulletinboard/database/ReadDataCallback.kt
@@ -1,7 +1,0 @@
-package com.ds_create.bulletinboard.database
-
-import com.ds_create.bulletinboard.data.Ad
-
-interface ReadDataCallback {
-    fun readData(list: List<Ad>)
-}
