+++ app/src/main/java/com/ds_create/bulletinboard/model/Ad.kt
@@ -1,0 +1,15 @@
+package com.ds_create.bulletinboard.model
+
+data class Ad(
+    val country: String? = null,
+    val city: String? = null,
+    val tel: String? = null,
+    val index: String? = null,
+    val withSend: String? = null,
+    val category: String? = null,
+    val title: String? = null,
+    val price: String? = null,
+    val description: String? = null,
+    val key: String? = null,
+    val uid: String? = null
+)
