+++ app/src/main/java/com/ds_create/bulletinboard/model/DbManager.kt
@@ -1,0 +1,46 @@
+package com.ds_create.bulletinboard.model
+
+
+import com.google.firebase.auth.ktx.auth
+import com.google.firebase.database.DataSnapshot
+import com.google.firebase.database.DatabaseError
+import com.google.firebase.database.ValueEventListener
+import com.google.firebase.database.ktx.database
+import com.google.firebase.ktx.Firebase
+
+class DbManager {
+    val db = Firebase.database.getReference("main")
+    val auth = Firebase.auth
+
+    fun publishAd(ad: Ad) {
+       if (auth.uid != null) {
+            db.child(ad.key ?: "empty")
+                .child(auth.uid!!)
+                .child("ad")
+                .setValue(ad)
+        }
+    }
+
+    fun readDataFromDb(readDataCallback: ReadDataCallback?) {
+        db.addListenerForSingleValueEvent(object : ValueEventListener {
+            override fun onDataChange(snapshot: DataSnapshot) {
+                val adArray = ArrayList<Ad>()
+                for (item in snapshot.children) {
+                    val ad = item.children.iterator().next().child("ad").getValue(Ad::class.java)
+                    if (ad != null) adArray.add(ad)
+                //                    ad?.let {
+//                        adArray.add(it)
+//                    }
+                }
+                readDataCallback?.readData(adArray)
+            }
+
+            override fun onCancelled(error: DatabaseError) {
+                TODO("Not yet implemented")
+            }
+        })
+    }
+    interface ReadDataCallback {
+        fun readData(list: ArrayList<Ad>)
+    }
+}
