+++ app/src/main/java/com/ds_create/bulletinboard/viewmodel/FirebaseViewModel.kt
@@ -1,0 +1,20 @@
+package com.ds_create.bulletinboard.viewmodel
+
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import com.ds_create.bulletinboard.model.Ad
+import com.ds_create.bulletinboard.model.DbManager
+
+class FirebaseViewModel: ViewModel() {
+
+    private val dbManager = DbManager()
+    val liveAdsData = MutableLiveData<ArrayList<Ad>>()
+
+    fun loadAllAds() {
+        dbManager.readDataFromDb(object: DbManager.ReadDataCallback {
+            override fun readData(list: ArrayList<Ad>) {
+                liveAdsData.value = list
+            }
+        })
+    }
+}
